+++ backend/canisters/user_index/impl/src/model/mod.rs
@@ -8,4 +8,3 @@ pub mod reported_messages;
 pub mod storage_index_user_sync_queue;
 pub mod user;
 pub mod user_map;
-pub mod user_referral_leaderboards;